makrx-store-backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication and security
pyjwt==2.8.0
cryptography==41.0.7
httpx==0.25.2

# Validation and serialization
pydantic[email]==2.5.0

# Storage and file handling
boto3==1.34.0
botocore==1.34.0

# Payments
stripe==7.8.0
razorpay==1.4.2

# Caching and queuing
redis==5.0.1
celery==5.3.4

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring and logging
structlog==23.2.0
